remove unwanted files from public directory
