+++ src/components/Display.js
@@ -1,0 +1,3 @@
+const Display = () => <div className="display">0</div>;
+
+export default Display;
